pyproject.toml
[tool.poetry]
authors = ["maicol1912"]
description = "Everything you need to know to get started with LangChain"
name = "langchain-exercises"
package-mode = false
version = "0.1.0"

[tool.poetry.dependencies]
bs4 = "^0.0.2"
chromadb = "^0.5.0"
firecrawl-py = "^0.0.13"
firestore = "^0.0.8"
langchain = "^0.3.0"
langchain-anthropic = "^0.2.0"
langchain-community = "^0.3.0"
langchain-deepseek = "^0.1.3"
langchain-google-firestore = "^0.3.0"
langchain-google-genai = "^2.0.0"
langchain-openai = "^0.3.0"
langchainhub = "^0.1.18"
python = ">=3.10.0,<3.12"
python-dotenv = "^1.0.1"
sentence-transformers = "^3.0.0"
tavily-python = "^0.3.3"
tiktoken = "^0.7.0"
wikipedia = "^1.4.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
exclude = [".cache"]
useLibraryCodeForTypes = true

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
ignore = ['W291', 'W292', 'W293']
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
